move test to shardingconfig
