+++ internal/configs/sharding/shardingconfig_test.go
@@ -1,6 +1,7 @@
 package shardingconfig
 
 import (
+	"fmt"
 	"math/big"
 	"testing"
 )
@@ -83,6 +84,29 @@ func TestCalcEpochNumber(t *testing.T) {
 		ep := MainnetSchedule.CalcEpochNumber(test.block)
 		if ep.Cmp(test.epoch) != 0 {
 			t.Errorf("CalcEpochNumber error: index %v, got %v, expect %v\n", i, ep, test.epoch)
+		}
+	}
+}
+
+func TestGetShardingStructure(t *testing.T) {
+	shardID := 0
+	numShard := 4
+	res := genShardingStructure(numShard, shardID, "http://s%d.t.hmy.io:9500", "ws://s%d.t.hmy.io:9800")
+	if len(res) != 4 || !res[0]["current"].(bool) || res[1]["current"].(bool) || res[2]["current"].(bool) || res[3]["current"].(bool) {
+		t.Error("Error when generating sharding structure")
+	}
+	for i := 0; i < numShard; i++ {
+		if res[i]["current"].(bool) != (i == shardID) {
+			t.Error("Error when generating sharding structure")
+		}
+		if res[i]["shardID"].(int) != i {
+			t.Error("Error when generating sharding structure")
+		}
+		if res[i]["http"].(string) != fmt.Sprintf("http://s%d.t.hmy.io:9500", i) {
+			t.Error("Error when generating sharding structure")
+		}
+		if res[i]["ws"].(string) != fmt.Sprintf("ws://s%d.t.hmy.io:9800", i) {
+			t.Error("Error when generating sharding structure")
 		}
 	}
 }
